fix: update Prisma schema to use Float instead of Decimal for SQLite compatibility

+++ apps/web/prisma/migrations/20240101000000_init/migration.sql
@@ -78,7 +78,7 @@ CREATE TABLE "categories" (
 -- CreateTable
 CREATE TABLE "transactions" (
     "id" TEXT NOT NULL PRIMARY KEY,
-    "amount" DECIMAL(12, 2) NOT NULL,
+    "amount" REAL NOT NULL,
     "kind" TEXT NOT NULL,
     "categoryId" TEXT NOT NULL,
     "note" TEXT,
